+++ src/containers/LoginPage/LoginPage.js
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import {authenticate} from 'src/store/actions/auth';
 
 
-export const Wrapper = styled.div`
+const Wrapper = styled.div`
   height: 100%;
   display: flex;
   align-items: center;
